Minor changes. Rename mic.py -> microphone.py.
